new atomiclong counter created. NearTxLocal and DhtTxLocal are set new counter instead old threadId
