+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/GridCacheMvcc.java
@@ -1188,7 +1188,7 @@ public final class GridCacheMvcc {
      * @return Removed candidate.
      */
     @Nullable public GridCacheMvccCandidate releaseLocal() {
-        return releaseLocal(Thread.currentThread().getId());
+        return releaseLocal(cctx.tm().getTxId());
     }
 
     /**
