+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/distributed/GridDistributedTxRemoteAdapter.java
@@ -159,7 +159,7 @@ public abstract class GridDistributedTxRemoteAdapter extends IgniteTxAdapter
             nodeId,
             xidVer,
             ctx.versions().last(),
-            Thread.currentThread().getId(),
+            ctx.tm().getTxId(),
             sys,
             plc,
             concurrency,
