+++ modules/core/src/main/java/org/apache/ignite/internal/processors/cache/distributed/near/GridNearTransactionalCache.java
@@ -524,7 +524,7 @@ public class GridNearTransactionalCache<K, V> extends GridNearCacheAdapter<K, V>
                         break; // While.
 
                     try {
-                        GridCacheMvccCandidate cand = entry.candidate(ctx.nodeId(), Thread.currentThread().getId());
+                        GridCacheMvccCandidate cand = entry.candidate(ctx.nodeId(), ctx.tm().getTxId());
 
                         AffinityTopologyVersion topVer = AffinityTopologyVersion.NONE;
